Update Refereneces to repo project name (#11)
I decided to name this repo "projects" because I found "node-monorepo" to be a little clunky. It also narrowed the scope of this repo to node-only projects, which I don't necessarily think is appropriate.

This commit just updates references to "node-monorepo" to "projects" so that it is no longer mentioned in here.